Fix the checking of units for compound models.
The issue was that the compound model called evaluate directly on all submodels without going through __call__ which means that prepare_inputs was being skipped. The unit validation was done inside prepare_units, so the fix is to separate out the validation of the units into a dedicated method. In addition, _model_evaluate_getter previously did not instantiate models when the compound model included a model class instead of an instance and evaluate was a class method - however we definitely need to always instantiate the model since e.g. input_units can be defined as an instance property.